feat(ai_hook): reduce coupling between verticals
