+++ .importlinter
@@ -10,7 +10,7 @@ type = layers
 layers = 
 	ggshield.__main__
 	ggshield.cmd.auth | ggshield.cmd.config | ggshield.cmd.hmsl | ggshield.cmd.honeytoken | ggshield.cmd.install | ggshield.cmd.plugin | ggshield.cmd.quota | ggshield.cmd.secret | ggshield.cmd.status | ggshield.cmd.utils
-	ggshield.verticals.auth | ggshield.verticals.hmsl | ggshield.verticals.secret
+	ggshield.verticals.ai | ggshield.verticals.auth | ggshield.verticals.hmsl | ggshield.verticals.secret
 	ggshield.core
 	click | ggshield.utils | pygitguardian
 ignore_imports = 
@@ -33,6 +33,7 @@ source_modules =
 	ggshield.cmd.status
 	ggshield.cmd.utils
 forbidden_modules = 
+	ggshield.verticals.ai
 	ggshield.verticals.auth
 	ggshield.verticals.hmsl
 	ggshield.verticals.secret
@@ -46,6 +47,7 @@ ignore_imports =
 	ggshield.cmd.hmsl.** -> ggshield.verticals.hmsl.**
 	ggshield.cmd.honeytoken.** -> ggshield.verticals.honeytoken
 	ggshield.cmd.honeytoken.** -> ggshield.verticals.honeytoken.**
+	ggshield.cmd.install -> ggshield.verticals.ai.installation
 	ggshield.cmd.install.** -> ggshield.verticals.install
 	ggshield.cmd.install.** -> ggshield.verticals.install.**
 	ggshield.cmd.plugin.** -> ggshield.core.plugin
@@ -54,6 +56,7 @@ ignore_imports =
 	ggshield.cmd.quota.** -> ggshield.verticals.quota.**
 	ggshield.cmd.secret.** -> ggshield.verticals.secret
 	ggshield.cmd.secret.** -> ggshield.verticals.secret.**
+	ggshield.cmd.secret.scan.ai_hook -> ggshield.verticals.ai.hooks
 	ggshield.cmd.status.** -> ggshield.verticals.status
 	ggshield.cmd.status.** -> ggshield.verticals.status.**
 	ggshield.cmd.utils.** -> ggshield.verticals.utils
