+++ ggshield/verticals/ai/__init__.py
@@ -1,0 +1,10 @@
+from .agents import AGENTS
+from .hooks import AIHookScanner
+from .installation import install_hooks
+
+
+__all__ = [
+    "AGENTS",
+    "AIHookScanner",
+    "install_hooks",
+]
